Improve seticons function

+++ .config/fish/functions/seticons.fish
@@ -13,17 +13,31 @@ function seticons
         set newname (string replace -r '\.[^.]+$' '.app' -- $basename)
 
         set appname "/Applications/$newname"
+
+        if not test -e "$appname"
+            echo "$appname not installed, skipping..."
+            continue
+        end
     
         if not fileicon test $appname
             sudo fileicon set $appname $file
             set haschanged 1
         end
     end
+
     if test $haschanged -eq 1
-        echo clearing dock icon cache
-        rm /var/folders/*/*/*/com.apple.dock.iconcache
-        echo clearing icon service cache
-        rm -r /var/folders/*/*/*/com.apple.iconservices*
+        set files /var/folders/*/*/*/com.apple.dock.iconcache
+        if test (count $files) -gt 0
+            echo clearing dock icon cache
+            rm /var/folders/*/*/*/com.apple.dock.iconcache
+        end
+
+        set files /var/folders/*/*/*/com.apple.iconservices*
+        if test (count $files) -gt 0
+            echo clearing icon service cache
+            rm -r /var/folders/*/*/*/com.apple.iconservices*
+        end
+
         echo restarting dock
         killall Dock
     end
